pyproject.toml
[project]
name = "agrismart"
version = "0.1.0"
description = "🐳 Agrismart: A Smart Farming Assistant for Sustainable Agriculture and Food Security"
readme = "README.md"
authors = [{ name = "Hinsun", email = "vanhoai.adv@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.5",
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "cloudinary>=1.44.1",
    "cryptography>=45.0.5",
    "facenet-pytorch>=2.6.0",
    "faiss-cpu>=1.11.0.post1",
    "faker>=37.5.3",
    "fastapi-camelcase>=2.0.0",
    "fastapi[standard]>=0.116.1",
    "firebase-admin>=7.0.0",
    "google-genai>=1.27.0",
    "hatchling>=1.27.0",
    "insightface>=0.7.3",
    "loguru>=0.7.3",
    "mediapipe>=0.10.21",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "python-socketio[asyncio-client]>=5.13.0",
    "starlette>=0.47.1",
    "supabase>=2.16.0",
    "supervision>=0.26.0",
    "tabulate>=0.9.0",
    "tensorflow-macos>=2.16.2; sys_platform == 'darwin'",
    "tensorflow>=2.16.2; sys_platform != 'darwin'",
    "ultralytics>=8.3.163",
]

[tool.black]
line-length = 120

[tool.uv.workspace]
members = [
    "packages/vision",
    "packages/core",
    "packages/domain",
    "packages/adapters",
]

[project.scripts]
agrismart = "agrismart:main"
benchmark = "benchmark:main"
training = "training:main"
dataset = "dataset:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
core = { workspace = true }
vision = { workspace = true }
domain = { workspace = true }
adapters = { workspace = true }
